Register `MenuViewComponent`
The `MenuViewComponent` offers a three-tier hierarchical list of the current root—typically `/Web`—and can be used to display a menu of the site. Optionally, this can be configured to display lower level tiers of navigation instead, if a full menu isn't necessary. This also necessitates registering a `HierarchicalTopicMappingService` in the navigation.