Establish stylesheet
Normally, we'd setup a stylesheet using SASS so it could be compiled and minimized. Given the nature of the project, however, we don't anticipate the stylesheet ever being used in production code; it is simply in place to provide temporary WYSIWYG previewing of the generated HTML. Given that, it can be a simple, static CSS page.