Establish OnTopic Editor for use locally
Establishes dependencies in `Startup` using the `AddTopicEditor()` extension method, and also registers the `/OnTopic` route using the `MapTopicEditorRoute()` extension method. In the `CanvasActivator`, handle both requests for the `EditorController` (which is associated with the `MapTopicEditorRoute()`), as well as the view components (which are used to display each individual attribute type, and which are handled collectively via the `StandardEditorComposer.ActivateEditorComponent()` method).