Conditionally add Razor runtime compilation in development
Razor Runtime Configuration allows Razor views (`*.cshtml`) to by dynamically recompiled when they're edited so that the entire application doesn't need to be reset each time a view changes. This greatly improves productivity when modifying the markup, but isn't used or necessary in production; as such, it should only be enabled in `IsDevelopment()` mode.