Establish custom style set for CkEditor
This allows the classes and styles displayed in the toolbar of the CkEditor plugin to be customized. Two sample classes are configured, but it is expected these will be customized based on the site's requirements. These can even be configured to inject inline styles instead of class names, which might be especially useful for this project.